Updated PerformantSocketServer to 2.0.0-beta1
Can now indicate if the connection should be left open and listening after a process, or gracefully closed.
Can now have a custom object in the settings config to pass around implementation details.
Can now have a custom state for a connection so that state can be passed around between processing calls.
Updated tracing information to contain more information for debugging if required.